@@ -1,45 +1,57 @@
-import {Component, OnInit, ViewChild} from "@angular/core";
-import {Observable} from "rxjs/Observable";
-import {Month} from "../shared/enum/Month";
-import {DayItem} from "../rest/model/DayItem";
-import {Person} from "../rest/model/Person";
-import {ActivatedRoute} from "@angular/router";
-import {PersonService} from "../shared/service/person.service";
-import {DayItemsContainer} from "../rest/model/DayItemsContainer";
-import * as json2csv from "json2csv";
-import {AttendanceService} from "../shared/service/attendance.service";
-import {HolidayComponent} from "./holiday/holiday.component";
+import {Component, OnDestroy, OnInit, ViewChild} from '@angular/core';
+import {Observable} from 'rxjs/Observable';
+import {Month} from '../shared/enum/Month';
+import {DayItem} from '../rest/model/DayItem';
+import {Person} from '../rest/model/Person';
+import {ActivatedRoute} from '@angular/router';
+import {PersonService} from '../shared/service/person.service';
+import {DayItemsContainer} from '../rest/model/DayItemsContainer';
+import * as json2csv from 'json2csv';
+import {AttendanceService} from '../shared/service/attendance.service';
+import {HolidayComponent} from './holiday/holiday.component';
+import {Subscription} from 'rxjs/Subscription';
 
 @Component({
   selector: 'emp-attendance',
-  templateUrl: './attendance.component.html',
-  providers: []
+  templateUrl: './attendance.component.html'
 })
-export class AttendanceComponent implements OnInit {
+export class AttendanceComponent implements OnInit, OnDestroy {
 
-  selectedMonth:Month;
-  selectedYear:number;
-  persons:Array<Person>;
-  dayItems:Array<DayItem>;
+  private getPersonsSub: Subscription;
+  private dayItemsSub: Subscription;
 
-  @ViewChild('holiday') holiday:HolidayComponent;
+  selectedMonth: Month;
+  selectedYear: number;
+  persons: Array<Person>;
+  dayItems: Array<DayItem>;
 
-  constructor(private personService: PersonService, private attendanceService:AttendanceService, private route:ActivatedRoute) {
-    let now:Date = new Date();
+  @ViewChild('holiday') holiday: HolidayComponent;
+
+  constructor(private personService: PersonService, private attendanceService: AttendanceService, private route: ActivatedRoute) {
+    let now: Date = new Date();
 
     this.selectedMonth = now.getMonth() + 1;
     this.selectedYear = now.getFullYear();
   }
 
   ngOnInit(): void {
-    this.personService.getPersons().subscribe(persons => {
+    this.getPersonsSub = this.personService.getPersons().subscribe(persons => {
       this.persons = persons;
       this._loadActualDayItems();
     });
   }
 
+  ngOnDestroy(): void {
+    if (this.dayItemsSub) {
+      this.dayItemsSub.unsubscribe();
+    }
+    if (this.getPersonsSub) {
+      this.getPersonsSub.unsubscribe();
+    }
+  }
+
   increaseMonth() {
-    if (this.selectedMonth == 12) {
+    if (this.selectedMonth === 12) {
       this.selectedMonth = 1;
       this.increaseYear();
     } else {
@@ -49,7 +61,7 @@ export class AttendanceComponent implements OnInit {
   }
 
   decreaseMonth() {
-    if (this.selectedMonth == 1) {
+    if (this.selectedMonth === 1) {
       this.selectedMonth = 12;
       this.decreaseYear();
     } else {
@@ -69,7 +81,7 @@ export class AttendanceComponent implements OnInit {
   }
 
   prefill() {
-    this._parseDayItemsObservable(this.personService.refillPerson(this.selectedMonth, this.selectedYear))
+    this._parseDayItemsObservable(this.personService.refillPerson(this.selectedMonth, this.selectedYear));
   }
 
   openHoliday() {
@@ -79,14 +91,14 @@ export class AttendanceComponent implements OnInit {
   exportIntoCsv() {
     let exportData = this._prepareExportData();
 
-    let blob = new Blob(["\ufffe", exportData], {type: 'application/csv'});
+    let blob = new Blob(['\ufffe', exportData], {type: 'application/csv'});
 
-    let a = window.document.createElement("a");
+    let a: HTMLAnchorElement = window.document.createElement('a');
     a.href = window.URL.createObjectURL(blob);
     a.download = 'export_' + this.selectedMonth + '_' + this.selectedYear + '.csv';
     document.body.appendChild(a);
     a.click();
-    document.body.removeChild(a)
+    document.body.removeChild(a);
   }
 
   _prepareExportData() {
@@ -99,24 +111,24 @@ export class AttendanceComponent implements OnInit {
     });
 
     let fields = ['date'];
-    for (var _i = 0; _i < this.persons.length; _i++) {
+    for (let _i = 0; _i < this.persons.length; _i++) {
       fields.push('data' + _i);
     }
 
     let data = [];
-    let totalDays:number =  new Date(this.selectedYear, this.selectedMonth, 0).getDate();
-    for (var dayIndex = 1; dayIndex <= totalDays; dayIndex++) {
-      let dataRow:Object = {'date': dayIndex + '.' + this.selectedMonth + '.'};
-      for (var personIndex = 0; personIndex < this.persons.length; personIndex++) {
-        let person:Person = this.persons[personIndex];
-        let dayItem:DayItem = this.attendanceService.findDayItem(this.dayItems, person, dayIndex, this.selectedMonth, this.selectedYear);
+    let totalDays: number =  new Date(this.selectedYear, this.selectedMonth, 0).getDate();
+    for (let dayIndex = 1; dayIndex <= totalDays; dayIndex++) {
+      let dataRow: Object = {'date': dayIndex + '.' + this.selectedMonth + '.'};
+      for (let personIndex = 0; personIndex < this.persons.length; personIndex++) {
+        let person: Person = this.persons[personIndex];
+        let dayItem: DayItem = this.attendanceService.findDayItem(this.dayItems, person, dayIndex, this.selectedMonth, this.selectedYear);
 
-        let dayItemValue:string = "";
+        let dayItemValue = '';
         if (dayItem && dayItem.recordSet && dayItem.recordSet.length > 0) {
           dayItem.recordSet.forEach(dayItemRecord => {
             dayItemValue += dayItemRecord.type.description;
             if (dayItemRecord.type.hoursType) {
-              dayItemValue += '(' + dayItemRecord.hoursCount + ')'
+              dayItemValue += '(' + dayItemRecord.hoursCount + ')';
             }
             dayItemValue += ',';
           });
@@ -132,8 +144,8 @@ export class AttendanceComponent implements OnInit {
     this._parseDayItemsObservable(this.personService.getDayItems(this.selectedMonth, this.selectedYear));
   }
 
-  _parseDayItemsObservable(dayItemsContainer:Observable<DayItemsContainer>) {
-    dayItemsContainer
+  _parseDayItemsObservable(dayItemsContainer: Observable<DayItemsContainer>) {
+    this.dayItemsSub = dayItemsContainer
       .subscribe(response => {
         this.dayItems = new Array();
 
