@@ -1,11 +1,10 @@
-import {Injectable} from "@angular/core";
-import {Observable} from "rxjs/Observable";
-import {PersonApi} from "../../rest/api/PersonApi";
-import {Person} from "../../rest/model/Person";
-import {DayItem} from "../../rest/model/DayItem";
-import {Month} from "../enum/Month";
-import {DayItemApi} from "../../rest/api/DayItemApi";
-import {DayItemsContainer} from "../../rest/model/DayItemsContainer";
+import {Injectable} from '@angular/core';
+import {Observable} from 'rxjs/Observable';
+import {PersonApi} from '../../rest/api/PersonApi';
+import {Person} from '../../rest/model/Person';
+import {Month} from '../enum/Month';
+import {DayItemApi} from '../../rest/api/DayItemApi';
+import {DayItemsContainer} from '../../rest/model/DayItemsContainer';
 
 @Injectable()
 export class PersonService {
@@ -18,11 +17,11 @@ export class PersonService {
     return this.personApi.personGet();
   }
 
-  getDayItems(month:Month, yearValue:number): Observable<DayItemsContainer> {
+  getDayItems(month: Month, yearValue: number): Observable<DayItemsContainer> {
     return this.dayItemApi.dayItemMonthYearGet(String(month), String(yearValue));
   }
 
-  refillPerson(month:Month, yearValue:number):Observable<DayItemsContainer> {
+  refillPerson(month: Month, yearValue: number): Observable<DayItemsContainer> {
     return this.dayItemApi.dayItemMonthYearRefillPut(String(month), String(yearValue));
   }
 
